requirements.txt
anthropic
speechrecognition
python-dotenv
requests
pydub
